backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
langgraph==0.0.26
langchain==0.1.12
langchain-core==0.1.32
langchain-openai==0.0.8
langchain-anthropic==0.1.4
langchain-mistralai
langchain-groq
langchain-google-genai
motor==3.3.2
pymongo==4.6.2
redis==5.0.3
httpx==0.27.0
python-dotenv==1.0.1
